Support single level reprojection
