+++ ndpyramid/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 
-from .core import pyramid_coarsen, pyramid_reproject
+from .coarsen import pyramid_coarsen
+from .reproject import pyramid_reproject
 from .regrid import pyramid_regrid
 from ._version import __version__
